requirements.txt
# Core dependencies for LatentVLA
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Vision and Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
einops>=0.6.0

# ML utilities
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Logging and Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Data processing
pandas>=2.0.0
h5py>=3.8.0
zarr>=2.14.0

# Configuration and utilities
pyyaml>=6.0
packaging>=23.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional GPU acceleration
# nvidia-ml-py3>=7.352.0

# Development tools (optional)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.4.0